pyproject.toml
[tool.poetry]
name = "my-api"
version = "0.1.0"
description = ""
authors = ["g42puts <gilmar.neo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.*"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic = "^2.10.4"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.1"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
testcontainers = "^4.9.0"
freezegun = "^1.5.1"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
tzdata = "^2024.2"
factory-boy = "^3.3.1"
psycopg = "^3.2.4"
snowflake-id = "^1.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev my_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=my_api -vv'
post_test = 'coverage html'
migrate = 'alembic revision --autogenerate'
migrate-apply = 'alembic upgrade head'
docker = 'docker compose -f docker-compose.yaml up -d'